livekit-examples/agent-demos/moviefone/requirements.txt
requests>=2.31.0
livekit
livekit-agents
livekit-plugins-openai
python-dotenv
numpy
tabulate>=0.9.0